REFACTORING:: 소설 list component화
