+++ package.json
@@ -19,7 +19,7 @@
     "react-dom": "^18.2.0",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.22.2",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "react-toastify": "^10.0.5",
     "recoil": "^0.7.7",
     "styled-components": "^6.1.8",
@@ -35,7 +35,7 @@
   "eslintConfig": {
     "extends": [
       "react-app",
-      "react-app/jest"
+      "plugin:react/recommended"
     ]
   },
   "browserslist": {
